move all folders under alert rule details
